adds a dummy implementation on windows
[#97325528]
