+++ cf-plugin/cmd/ssh_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package cmd_test
 
 import (
