+++ cmd/ssh-proxy/main_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package main_test
 
 import (
