+++ cmd/sshd/main_not_windoz_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package main_test
 
 import (
