+++ handlers/session_channel_handler_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package handlers_test
 
 import (
