+++ handlers/signals_not_windoz.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package handlers
 
 import (
