+++ handlers/termcodes_attr_setter_windows.go
@@ -1,0 +1,5 @@
+// +build windows
+
+package handlers
+
+var TermAttrSetters map[uint8]TCSetter = map[uint8]TCSetter{}
