+++ handlers/termcodes_termios_not_windoz.go
@@ -1,0 +1,7 @@
+// +build !windows
+
+package handlers
+
+import "syscall"
+
+type Termios syscall.Termios
