+++ scp/atime/atime_windows.go
@@ -1,0 +1,12 @@
+// +build windows
+
+package atime
+
+import (
+	"os"
+	"syscall"
+)
+
+func accessTimespec(fileInfo os.FileInfo) syscall.Timespec {
+	return syscall.Timespec{}
+}
