+++ scp/scp_suite_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package scp_test
 
 import (
